# internal/publicoas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package publicoas

import (
	"fmt"
)

func (s *DefaultErrorStatusCodeWithHeaders) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// DefaultErrorStatusCodeWithHeaders wraps ErrorResponse with status code and response headers.
type DefaultErrorStatusCodeWithHeaders struct {
	StatusCode int
	XRequestID OptString
	Response   ErrorResponse
}

// GetStatusCode returns the value of StatusCode.
func (s *DefaultErrorStatusCodeWithHeaders) GetStatusCode() int {
	return s.StatusCode
}

// GetXRequestID returns the value of XRequestID.
func (s *DefaultErrorStatusCodeWithHeaders) GetXRequestID() OptString {
	return s.XRequestID
}

// GetResponse returns the value of Response.
func (s *DefaultErrorStatusCodeWithHeaders) GetResponse() ErrorResponse {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *DefaultErrorStatusCodeWithHeaders) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetXRequestID sets the value of XRequestID.
func (s *DefaultErrorStatusCodeWithHeaders) SetXRequestID(val OptString) {
	s.XRequestID = val
}

// SetResponse sets the value of Response.
func (s *DefaultErrorStatusCodeWithHeaders) SetResponse(val ErrorResponse) {
	s.Response = val
}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Code      string    `json:"code"`
	Message   string    `json:"message"`
	RequestId OptString `json:"requestId"`
}

// GetCode returns the value of Code.
func (s *ErrorResponse) GetCode() string {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() string {
	return s.Message
}

// GetRequestId returns the value of RequestId.
func (s *ErrorResponse) GetRequestId() OptString {
	return s.RequestId
}

// SetCode sets the value of Code.
func (s *ErrorResponse) SetCode(val string) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val string) {
	s.Message = val
}

// SetRequestId sets the value of RequestId.
func (s *ErrorResponse) SetRequestId(val OptString) {
	s.RequestId = val
}

// Ref: #/components/schemas/HealthResponse
type HealthResponse struct {
	Status string `json:"status"`
}

// GetStatus returns the value of Status.
func (s *HealthResponse) GetStatus() string {
	return s.Status
}

// SetStatus sets the value of Status.
func (s *HealthResponse) SetStatus(val string) {
	s.Status = val
}

// HealthResponseHeaders wraps HealthResponse with response headers.
type HealthResponseHeaders struct {
	XRequestID OptString
	Response   HealthResponse
}

// GetXRequestID returns the value of XRequestID.
func (s *HealthResponseHeaders) GetXRequestID() OptString {
	return s.XRequestID
}

// GetResponse returns the value of Response.
func (s *HealthResponseHeaders) GetResponse() HealthResponse {
	return s.Response
}

// SetXRequestID sets the value of XRequestID.
func (s *HealthResponseHeaders) SetXRequestID(val OptString) {
	s.XRequestID = val
}

// SetResponse sets the value of Response.
func (s *HealthResponseHeaders) SetResponse(val HealthResponse) {
	s.Response = val
}

func (*HealthResponseHeaders) getHealthzRes() {}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// ServiceUnavailableErrorHeaders wraps ErrorResponse with response headers.
type ServiceUnavailableErrorHeaders struct {
	XRequestID OptString
	Response   ErrorResponse
}

// GetXRequestID returns the value of XRequestID.
func (s *ServiceUnavailableErrorHeaders) GetXRequestID() OptString {
	return s.XRequestID
}

// GetResponse returns the value of Response.
func (s *ServiceUnavailableErrorHeaders) GetResponse() ErrorResponse {
	return s.Response
}

// SetXRequestID sets the value of XRequestID.
func (s *ServiceUnavailableErrorHeaders) SetXRequestID(val OptString) {
	s.XRequestID = val
}

// SetResponse sets the value of Response.
func (s *ServiceUnavailableErrorHeaders) SetResponse(val ErrorResponse) {
	s.Response = val
}

func (*ServiceUnavailableErrorHeaders) getHealthzRes() {}

// TooManyRequestsErrorHeaders wraps ErrorResponse with response headers.
type TooManyRequestsErrorHeaders struct {
	Ratelimit       OptString
	RatelimitPolicy OptString
	RetryAfter      OptString
	XRequestID      OptString
	Response        ErrorResponse
}

// GetRatelimit returns the value of Ratelimit.
func (s *TooManyRequestsErrorHeaders) GetRatelimit() OptString {
	return s.Ratelimit
}

// GetRatelimitPolicy returns the value of RatelimitPolicy.
func (s *TooManyRequestsErrorHeaders) GetRatelimitPolicy() OptString {
	return s.RatelimitPolicy
}

// GetRetryAfter returns the value of RetryAfter.
func (s *TooManyRequestsErrorHeaders) GetRetryAfter() OptString {
	return s.RetryAfter
}

// GetXRequestID returns the value of XRequestID.
func (s *TooManyRequestsErrorHeaders) GetXRequestID() OptString {
	return s.XRequestID
}

// GetResponse returns the value of Response.
func (s *TooManyRequestsErrorHeaders) GetResponse() ErrorResponse {
	return s.Response
}

// SetRatelimit sets the value of Ratelimit.
func (s *TooManyRequestsErrorHeaders) SetRatelimit(val OptString) {
	s.Ratelimit = val
}

// SetRatelimitPolicy sets the value of RatelimitPolicy.
func (s *TooManyRequestsErrorHeaders) SetRatelimitPolicy(val OptString) {
	s.RatelimitPolicy = val
}

// SetRetryAfter sets the value of RetryAfter.
func (s *TooManyRequestsErrorHeaders) SetRetryAfter(val OptString) {
	s.RetryAfter = val
}

// SetXRequestID sets the value of XRequestID.
func (s *TooManyRequestsErrorHeaders) SetXRequestID(val OptString) {
	s.XRequestID = val
}

// SetResponse sets the value of Response.
func (s *TooManyRequestsErrorHeaders) SetResponse(val ErrorResponse) {
	s.Response = val
}

func (*TooManyRequestsErrorHeaders) getHealthzRes() {}